// calc-front/src/@vex/interfaces/navigation-item.interface.ts
import { Icon } from '@visurel/iconify-angular';

export type NavigationItem = NavigationLink | NavigationDropdown | NavigationSubheading;

export interface NavigationLink {
  type: 'link';
  route: string | any;
  fragment?: string;
  label: string;
  icon?: Icon;
  routerLinkActive?: { exact: boolean };
  badge?: {
    value: string;
    background: string;
    color: string;
  };
}

export interface NavigationDropdown {
  type: 'dropdown';
  label: string;
  icon?: Icon;
  children: Array<NavigationLink | NavigationDropdown>;
  badge?: {
    value: string;
    background: string;
    color: string;
  };
}

export interface NavigationSubheading {
  type: 'subheading';
  label: string;
  isMega?: boolean;
  children: Array<NavigationLink | NavigationDropdown>;
}